machine_learning_app/requirements.txt
streamlit
numpy
matplotlib
scikit-learn
pillow
requests
streamlit-lottie